common_event_recorder

Module recorder

Structs

EventHandlerOptions

EventHandlerOptions is the options for the event handler.

EventProcessor 🔒

EventRecorderImpl

Implementation of EventRecorder that records the events and processes them in the background by the EventHandler.

EventRecorderOptions

Configuration options for the event recorder.

Constants

DEFAULT_BUFFER_SIZE 🔒

DEFAULT_CHANNEL_SIZE 🔒

DEFAULT_COMPACTION_TIME_WINDOW

The default compaction time window for the events table.

DEFAULT_EVENTS_TABLE_NAME

The default table name for storing the events.

DEFAULT_EVENTS_TABLE_TTL 🔒

The default TTL(90 days) for the events table.

DEFAULT_FLUSH_INTERVAL_SECONDS

The time interval for flushing batched events to the event handler.

DEFAULT_MAX_RETRY_TIMES 🔒

EVENTS_TABLE_PAYLOAD_COLUMN_NAME

The column name for the event payload.

EVENTS_TABLE_TIMESTAMP_COLUMN_NAME

The column name for the event timestamp.

EVENTS_TABLE_TYPE_COLUMN_NAME

The column name for the event type.

Traits

Event

Event trait defines the interface for events that can be recorded and persisted as the system table. By default, the event will be persisted as the system table with the following schema:

EventHandler

EventHandler trait defines the interface for how to handle the event.

EventRecorder

EventRecorder trait defines the interface for recording events.

Eventable

Eventable trait defines the interface for objects that can be converted to Event.

Functions

build_row_inserts_request

Builds the row inserts request for the events that will be persisted to the events table. The events should have the same event type, or it will return an error.

group_events_by_type

Groups events by its event_type.

validate_events 🔒

Type Aliases

EventRecorderRef

EventRecorderRef is the reference to the event recorder.